+++ ownCloudSDK/Bookmark/OCBookmark.h
@@ -43,6 +43,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(readonly,nullable) NSString *userName; //!< Convenience method for accessing the userName stored in the authenticationData
 
+@property(strong,nullable) NSString *userDisplayName; //!< Display name of a user
+
 @property(strong,nullable) NSURL *originURL; //!< URL originally provided by the user, which then redirected to .url. In case .url becomes invalid, the originURL can be used to find the new server. If originURL is set, UI should present it prominently - while also displaying .url near it.
 
 @property(strong,nullable) OCCertificate *certificate; //!< Certificate last used by the server this bookmark refers to